Run libfake as a library
Previously I used fedora and microdnf to install libfake,
but it is fully awalible to build it on the fly
with just gcc.

I got a problem while trying to reuse libfake build stage to run an jar with it. So I separate it and copy lib so from the stage